pyproject.toml
[build-system]
requires = ["poetry-core>=1.2.0b2"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "retayn"
version = "0.1.0"
description = "Insert Description"
license = "MIT license"
authors = ["Jonas Teufel <jonseb1998@gmail.com>"]
maintainers = ["Jonas Teufel <jonseb1998@gmail.com>"]
readme = "README.rst"
packages = [
    { include = "retayn" },
]
include = [
    "retayn/templates",
    "retayn/VERSION"
]

[tool.poetry.scripts]
retayn = 'retayn.cli:cli'

[tool.poetry.dependencies]
python = ">=3.8.0,<4.0"
click = ">=7.1.2"
jinja2 = ">=3.1.2"
pyyaml = ">=6.0"

[tool.poetry.dev-dependencies]
pytest = ">=7.1.3"
sphinx = ">=5.2.0"
sphinx-rtd-theme = ">=1.0.0"
virtualenv = ">=20.10.0"
poetry-bumpversion = ">=0.2.0"
python-decouple = ">=3.6"

# BUMPVERSION
[tool.poetry_bumpversion.file."pycomex/__init__.py"]
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[tool.poetry_bumpversion.file."pycomex/VERSION"]
search = "{current_version}"
replace = "{new_version}"

[tool.poetry_bumpversion.file."README.rst"]
search = "{current_version}"
replace = "{new_version}"